Merge branch 'dev' of https://github.com/Elronnd/Spelunk-1
into Elronnd-dev

Elronnd has very graciously made several changes to the source code to
make it compliant with D standards

+++ src/sym.d
@@ -29,7 +29,7 @@ struct symbol
   // images (possibly including sprites in the future
 }
 
-symbol symdata( char character, ushort effects )
+symbol symdata( char character, ulong effects )
 {
   static if( !TEXT_EFFECTS )
   { ushort color = effects;
